api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
python-multipart==0.0.20
PyJWT==2.10.1
passlib==1.7.4
python-dotenv==1.1.0
pymongo==4.13.2
rich
